requirements.txt
streamlit==1.42.0
pandas==2.2.3
plotly==6.0.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.13.3
rich==13.9.4
